backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.2.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
